Add files via upload
In this project, we address a common issue with email management: the overwhelming number of emails that make it difficult to read each one in detail. To solve this problem, we propose a "Text to Voice" feature. With this solution, the content of each email will be automatically read aloud, allowing users to listen to the complete text without having to read it.